refactor: update usage
