+++ mcpauth/__init__.py
@@ -1,9 +1,9 @@
 import logging
-from typing import Any, Literal, Union
+from typing import List, Literal, Optional, Union
 
-from .middleware.create_bearer_auth import BaseBearerAuthConfig, BearerAuthConfig
+from .middleware.create_bearer_auth import BearerAuthConfig
 from .types import VerifyAccessTokenFunction
-from .config import MCPAuthConfig
+from .config import AuthServerConfig
 from .exceptions import MCPAuthAuthServerException, AuthServerExceptionCode
 from .utils import validate_server_config
 from starlette.middleware.base import BaseHTTPMiddleware
@@ -16,12 +16,14 @@ class MCPAuth:
     functions for handling OAuth 2.0-related tasks and bearer token auth.
 
     See Also: https://mcp-auth.dev for more information about the library and its usage.
-
-    :param config: An instance of `MCPAuthConfig` containing the server configuration.
     """
 
-    def __init__(self, config: MCPAuthConfig):
-        result = validate_server_config(config.server)
+    def __init__(self, server: AuthServerConfig):
+        """
+        :param server: Configuration for the remote authorization server.
+        """
+
+        result = validate_server_config(server)
 
         if not result.is_valid:
             logging.error(
@@ -37,13 +39,13 @@ class MCPAuth:
             for warning in result.warnings:
                 logging.warning(f"- {warning}")
 
-        self.config = config
+        self.server = server
 
     def metadata_response(self) -> JSONResponse:
         """
         Returns a response containing the server metadata in JSON format with CORS support.
         """
-        server_config = self.config.server
+        server_config = self.server
 
         response = JSONResponse(
             server_config.metadata.model_dump(exclude_none=True),
@@ -56,22 +58,26 @@ class MCPAuth:
     def bearer_auth_middleware(
         self,
         mode_or_verify: Union[Literal["jwt"], VerifyAccessTokenFunction],
-        config: BaseBearerAuthConfig = BaseBearerAuthConfig(),
-        jwt_options: dict[str, Any] = {},
+        audience: Optional[str] = None,
+        required_scopes: Optional[List[str]] = None,
+        show_error_details: bool = False,
+        leeway: float = 60,
     ) -> type[BaseHTTPMiddleware]:
         """
         Creates a middleware that handles bearer token authentication.
 
         :param mode_or_verify: If "jwt", uses built-in JWT verification; or a custom function that
         takes a string token and returns an `AuthInfo` object.
-        :param config: Configuration for the Bearer auth handler, including audience, required
-        scopes, etc.
-        :param jwt_options: Optional dictionary of additional options for JWT verification
-        (`jwt.decode`). Not used if a custom function is provided.
+        :param audience: Optional audience to verify against the token.
+        :param required_scopes: Optional list of scopes that the token must contain.
+        :param show_error_details: Whether to include detailed error information in the response.
+        Defaults to `False`.
+        :param leeway: Optional leeway in seconds for JWT verification (`jwt.decode`). Defaults to
+        `60`. Not used if a custom function is provided.
         :return: A middleware class that can be used in a Starlette or FastAPI application.
         """
 
-        metadata = self.config.server.metadata
+        metadata = self.server.metadata
         if isinstance(mode_or_verify, str) and mode_or_verify == "jwt":
             from .utils import create_verify_jwt
 
@@ -82,7 +88,7 @@ class MCPAuth:
 
             verify = create_verify_jwt(
                 metadata.jwks_uri,
-                options=jwt_options,
+                leeway=leeway,
             )
         elif callable(mode_or_verify):
             verify = mode_or_verify
@@ -94,5 +100,11 @@ class MCPAuth:
         from .middleware.create_bearer_auth import create_bearer_auth
 
         return create_bearer_auth(
-            verify, BearerAuthConfig(issuer=metadata.issuer, **config.model_dump())
+            verify,
+            BearerAuthConfig(
+                issuer=metadata.issuer,
+                audience=audience,
+                required_scopes=required_scopes,
+                show_error_details=show_error_details,
+            ),
         )
