--- mcpauth/models/__init__.py
@@ -1,7 +1,0 @@
-from .auth_server import (
-    AuthServerConfig as AuthServerConfig,
-    AuthServerType as AuthServerType,
-)
-from .oauth import (
-    AuthorizationServerMetadata as AuthorizationServerMetadata,
-)
