+++ mcpauth/utils/__init__.py
@@ -2,7 +2,6 @@ from ._create_verify_jwt import create_verify_jwt as create_verify_jwt
 from ._fetch_server_config import (
     fetch_server_config as fetch_server_config,
     fetch_server_config_by_well_known_url as fetch_server_config_by_well_known_url,
-    ServerMetadataPaths as ServerMetadataPaths,
 )
 from ._validate_server_config import (
     validate_server_config as validate_server_config,
