+++ mcpauth/utils/_create_verify_jwt.py
@@ -1,4 +1,4 @@
-from typing import Annotated, Any, List, Optional, Union
+from typing import Annotated, List, Optional, Union
 from jwt import PyJWK, PyJWKClient, PyJWTError, decode
 from pydantic import BaseModel, StringConstraints, ValidationError
 from ..types import AuthInfo, VerifyAccessTokenFunction
@@ -23,8 +23,7 @@ class JwtBaseModel(BaseModel):
 def create_verify_jwt(
     input: Union[str, PyJWKClient, PyJWK],
     algorithms: List[str] = ["RS256", "PS256", "ES256", "ES384", "ES512"],
-    leeway: int = 60,
-    options: dict[str, Any] = {},
+    leeway: float = 60,
 ) -> VerifyAccessTokenFunction:
     """
     Creates a JWT verification function using the provided JWKS URI.
@@ -35,7 +34,6 @@ def create_verify_jwt(
         - An instance of `PyJWK` that represents a single JWK.
     :param algorithms: A list of acceptable algorithms for verifying the JWT signature.
     :param leeway: The amount of leeway (in seconds) to allow when checking the expiration time of the JWT.
-    :param options: Additional options to pass to the JWT decode function (`jwt.decode`).
     :return: A function that can be used to verify JWTs.
     """
 
@@ -66,8 +64,7 @@ def create_verify_jwt(
                 options={
                     "verify_aud": False,
                     "verify_iss": False,
-                }
-                | options,
+                },
             )
             base_model = JwtBaseModel(**decoded)
             scopes = base_model.scope or base_model.scopes
