+++ mcpauth/utils/_validate_server_config.py
@@ -1,7 +1,7 @@
 from enum import Enum
 from typing import Any, Dict, List, Optional
 from pydantic import BaseModel, ValidationError
-from mcpauth.models.auth_server import AuthServerConfig
+from ..config import AuthServerConfig
 
 
 class AuthServerConfigErrorCode(str, Enum):
