+++ tests/utils/validate_server_config_test.py
@@ -1,5 +1,4 @@
-from mcpauth.models.auth_server import AuthServerConfig, AuthServerType
-from mcpauth.models.oauth import AuthorizationServerMetadata
+from mcpauth.config import AuthServerConfig, AuthServerType, AuthorizationServerMetadata
 from mcpauth.utils import (
     validate_server_config,
     AuthServerConfigErrorCode,
